requirements.txt
accelerate==1.0.1
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
alembic==1.13.3
aniso8601==9.0.1
annotated-types==0.7.0
async-timeout==4.0.3
attrs==23.2.0
black==24.10.0
blinker==1.8.2
boto3==1.35.38
botocore==1.35.38
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
cloudpickle==2.2.1
contourpy==1.3.0
cycler==0.12.1
databricks-sdk==0.34.0
datasets
Deprecated==1.2.14
dill
docker==7.1.0
docker-pycreds==0.4.0
exceptiongroup==1.2.2
filelock==3.16.1
fire==0.7.0
Flask==3.0.3
fonttools==4.54.1
frozenlist==1.4.1
fsspec==2024.6.1
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.35.0
google-pasta==0.2.0
graphene==3.3
graphql-core==3.2.4
graphql-relay==3.2.0
gunicorn==23.0.0
huggingface-hub==0.25.2
idna==3.10
importlib-metadata==6.11.0
iniconfig==2.0.0
isort==5.13.2
itsdangerous==2.2.0
jaxtyping==0.2.34
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.7
Mako==1.3.5
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.1
matplotlib==3.9.2
mdurl==0.1.2
mlflow==2.16.2
mlflow-skinny==2.16.2
mock==4.0.3
mpmath==1.3.0
multidict==6.1.0
multiprocess
mypy-extensions==1.0.0
networkx==3.4
numpy==1.26.4
opentelemetry-api==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
packaging==24.1
pandas==2.2.3
pathos
pathspec==0.12.1
pillow==10.4.0
platformdirs==4.3.6
pluggy==1.5.0
pox==0.3.5
ppft==1.7.6.9
propcache==0.2.0
protobuf==4.25.5
psutil==6.0.0
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
pyparsing==3.1.4
pytest==8.3.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
rich==13.9.2
rpds-py==0.20.0
rsa==4.9
s3transfer==0.10.3
safetensors==0.4.5
sagemaker==2.232.2
sagemaker-core==1.0.10
sagemaker-mlflow==0.1.0
schema==0.7.7
scikit-learn==1.5.2
scipy==1.14.1
sentry-sdk==2.16.0
setproctitle==1.3.3
six==1.16.0
smdebug-rulesconfig==1.0.1
smmap==5.0.1
SQLAlchemy==2.0.35
sqlparse==0.5.1
sympy==1.13.3
tblib==3.0.0
termcolor==2.5.0
threadpoolctl==3.5.0
tokenizers==0.20.1
tomli==2.0.2
torch==2.4.1
tqdm==4.66.5
transformers==4.45.2
typeguard==2.13.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wandb==0.18.3
Werkzeug==3.0.4
wrapt==1.16.0
xxhash==3.5.0
yarl==1.14.0
zipp==3.20.2
